refactor(logger): moved logger to aurelia-form.js
