+++ src/aurelia-form.js
@@ -76,3 +76,8 @@ export function configure(aurelia, configCallback) {
 }
 
 export {Config};
+
+/* aurelia-form logger */
+import {getLogger} from 'aurelia-logging';
+const logger = getLogger('aurelia-form');
+export {logger};
