--- src/logger.js
@@ -1,3 +1,0 @@
-import {getLogger} from 'aurelia-logging';
-const logger = getLogger('aurelia-form');
-export {logger};
